mysite/requirements.txt
asgiref==3.7.2
autopep8==2.3.1
Django==5.0.9
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
packaging==24.2
pillow==11.0.0
pycodestyle==2.12.1
sqlparse==0.4.4
typing_extensions==4.9.0
